tecdsa.PublicParameters extracted to a separate file
Code wasn't modified. It was just a cut&paste to a new file.
